Split batching submission config reader from the config's definition.
