+++ ledger/participant-state/kvutils/src/main/scala/com/daml/ledger/participant/state/kvutils/api/BatchingLedgerWriterConfig.scala
@@ -1,9 +1,7 @@
 // Copyright (c) 2020 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.participant.state.kvutils.app.batch
-
-import scopt.Read
+package com.daml.ledger.participant.state.kvutils.api
 
 import scala.concurrent.duration._
 
@@ -32,20 +30,4 @@ object BatchingLedgerWriterConfig {
       maxBatchWaitDuration = 100.millis,
       maxBatchConcurrentCommits = 5
     )
-
-  implicit val optionsReader: Read[BatchingLedgerWriterConfig] = Read.mapRead[String, String].map {
-    _.foldLeft[BatchingLedgerWriterConfig](reasonableDefault) {
-      case (config, (key, value)) =>
-        key match {
-          case "enable" => config.copy(enableBatching = Read.booleanRead.reads(value))
-          case "max-queue-size" => config.copy(maxBatchQueueSize = Read.intRead.reads(value))
-          case "max-batch-size-bytes" => config.copy(maxBatchSizeBytes = Read.longRead.reads(value))
-          case "max-wait-millis" =>
-            config.copy(maxBatchWaitDuration = Duration(Read.longRead.reads(value), MILLISECONDS))
-          case "max-concurrent-commits" =>
-            config.copy(maxBatchConcurrentCommits = Read.intRead.reads(value))
-          case _ => config
-        }
-    }
-  }
 }
