+++ ledger/participant-state/kvutils/src/test/suite/scala/com/daml/ledger/validator/LogAppendingCommitStrategySpec.scala
@@ -79,4 +79,3 @@ class LogAppendingCommitStrategySpec extends AsyncWordSpec with Matchers with Mo
 
   private val aStateValue: DamlStateValue = DamlStateValue.getDefaultInstance
 }
-
